Rename: Firebase API 폴더 내 파일 명 dummyDataSetter, dataGetter로 변경
